modify hvm_pci_register to use uint16_t for domain instead of uint8_t

+++ xen/arch/x86/hvm/pci_emul.c
@@ -64,7 +64,7 @@ end_handle:
 }
 
 int hvm_register_pcidev(domid_t domid, ioservid_t id,
-                        uint8_t domain, uint8_t bus,
+                        uint16_t domain, uint8_t bus,
                         uint8_t device, uint8_t function)
 {
     struct domain *d;
